// ebegu-web/src/hybrid/gesuch/betreuung/pendenz/pendenzen-betreuungen-filter.interface.ts
export interface DVPendenzenBetreuungenFilter {
    betreuungsNummer?: string;
    gemeinde?: string;
    name?: string;
    vorname?: string;
    geburtsdatum?: string;
    antragTyp?: string;
    periodenString?: string;
    eingangsdatum?: string;
    angebote?: string;
    institutionen?: string;
}